#6
At the beginning of a chapter, pressing BACK will do nothing; press FORWARD will display the next part of the chapter
-- DONE, tested.

When a book is loaded, the first part of the chapter will be transmitted.
-- DONE, tested

At the end of a chapter, pressing BACK will display the previous part of the chapter; pressing FORWARD will display the next chapter.
-- TODO