#18
If multiple request as received, the next chapter chunk will not be prepared until the "RX: MSG_TYPE_ACK Rx: no error" msg is received, only the chunk index is incremented.
-- DONE. Not tested yet


If the index reach the two extremes of a chapter, it will stick there without going to the next chapter.
-- Change this a bit. No need for the sticking point. Just go to the adjacent chapter.